scripts:pkgs:win: add package installation script for Windows
This will install the toolchain and libraries needed to build xNVMe in a
manner compatible with fio. As well as the toolchains and libraries
needed to build fio itself.

The chocolatey package manager is used for installing the software.

This change moves the 'where' check to after the various installation
commands, thereby enabling the script to also upgrade packages. E.g. if
somebody has an old version of the mingw/msys without the ioctl-headers,
then re-running the batch-script will upgrade it.

Also, this presents issues where e.g. GitHUB might have another version
installed of tools that we need. The 'where' check is thus there to
determine if after running the commands, that the tool is available or
not.

The script is named to match the naming convention of Windows on GitHUB
actions.

Signed-off-by: Rishabh Shukla <rishabh.sh@samsung.com>
Signed-off-by: Michael Bang <mi.bang@samsung.com>
Signed-off-by: Simon A. F. Lund <simon.lund@samsung.com>